refactor: 공통 컴포넌트 패키지 이동
