+++ app/src/androidTest/java/woowacourse/payments/ui/component/CardNumberTextFieldTest.kt
@@ -1,7 +1,9 @@
-package woowacourse.payments.cardaddition.component
+package woowacourse.payments.ui.component
 
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
@@ -10,38 +12,34 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class ExpiredDateTextFieldTest {
+class CardNumberTextFieldTest {
     @get:Rule
     val composeRule = createComposeRule()
 
     @Before
     fun setUp() {
         composeRule.setContent {
-            val (expiredDate: String, setExpiredDate: (String) -> Unit) =
-                remember {
-                    mutableStateOf(
-                        "",
-                    )
-                }
+            val (cardNumber: String, setCardNumber: (String) -> Unit) = remember { mutableStateOf("") }
 
-            ExpiredDateTextField(
-                value = expiredDate,
-                onValueChange = setExpiredDate,
+            CardNumberTextField(
+                value = cardNumber,
+                onValueChange = setCardNumber,
                 isError = false,
+                modifier = Modifier.fillMaxWidth(),
             )
         }
     }
 
     @Test
-    fun `만료일의_경우_입력할_때_자동으로_기호가_삽입된다`() {
+    fun `카드_번호의_경우_입력할_때_자동으로_기호가_삽입된다`() {
         // when
         composeRule
             .onNodeWithText("")
-            .performTextInput("1225")
+            .performTextInput("1234567812345678")
 
         // then
         composeRule
-            .onNodeWithText("12 / 25")
+            .onNodeWithText("1234 - 5678 - 1234 - 5678")
             .assertIsDisplayed()
     }
 }
